refactor: remove workout exercise using workoutId and exerciseId

+++ src/use-cases/workouts/remove-exercise-from-workout.use-case.ts
@@ -1,22 +1,39 @@
 import { ResourceNotFoundError } from '@/errors/resource-not-found.error'
 import { WorkoutExercisesRepository } from '@/repositories/workout-exercises.repository'
+import { WorkoutsRepository } from '@/repositories/workouts.repository'
 
 interface RemoveExerciseFromWorkoutUseCaseRequest {
-  workoutExerciseId: string
+  workoutId: string
+  exerciseId: string
 }
 
 export class RemoveExerciseFromWorkoutUseCase {
-  constructor(private workoutExercisesRepository: WorkoutExercisesRepository) {}
+  constructor(
+    private workoutsRepository: WorkoutsRepository,
+    private workoutExercisesRepository: WorkoutExercisesRepository,
+  ) {}
 
   async execute({
-    workoutExerciseId,
+    workoutId,
+    exerciseId,
   }: RemoveExerciseFromWorkoutUseCaseRequest): Promise<void> {
-    try {
-      await this.workoutExercisesRepository.removeExerciseFromWorkout(
-        workoutExerciseId,
-      )
-    } catch {
-      throw new ResourceNotFoundError('Exercise not found.')
+    const workout = await this.workoutsRepository.findById(workoutId)
+    if (!workout) {
+      throw new ResourceNotFoundError('Workout not found.')
     }
+
+    const workoutExercise =
+      await this.workoutExercisesRepository.findByWorkoutAndExercise(
+        workoutId,
+        exerciseId,
+      )
+
+    if (!workoutExercise) {
+      throw new ResourceNotFoundError('Exercise not found in this workout.')
+    }
+
+    await this.workoutExercisesRepository.removeExerciseFromWorkout(
+      workoutExercise.id,
+    )
   }
 }
